requirements.txt
asgiref==3.7.2
black==24.1.1
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
python-memcached==1.62
pytz==2024.1
redis==5.0.1
sqlparse==0.4.4
